Add files via upload
The R program downloads TCGA-BRCA (Gencode 36) RNA-Seq and copy number variation data from GDC using TCGAbiolinks. Users can edit the gene name and/or the subset selection in a box. The gene name should be in the HGNC ID to identify the gene. The program displays basic graphs and runs a survival analysis on the selected gene. The survival analysis identifies the optimal cut-off point that minimizes the p-value of the Log Rank Test and draws the Kaplan-Meier (KM) curve at the point. The program checks the overall trend of the difference in survival by drawing KM curves at the lower/median/upper quartile. 